+++ vector/src/vector_types/mod.rs
@@ -20,7 +20,6 @@ mod real;
 pub use self::real::*;
 mod time_freq;
 pub use self::time_freq::*;
-#[cfg(feature="std")]
 pub mod combined_ops;
 mod rededicate_and_relations;
 pub use self::rededicate_and_relations::*;